Feature/purchase#1

+++ src/main/java/javachip/service/impl/ProductServiceImpl.java
@@ -7,21 +7,25 @@
 
 package javachip.service.impl;
 
+import jakarta.transaction.Transactional;
 import javachip.dto.ProductDto;
 import javachip.entity.Product;
+import javachip.repository.ProductImageRepository;
 import javachip.repository.ProductRepository;
 import javachip.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class ProductServiceImpl implements ProductService {
 
     private final ProductRepository repository;
+    private final ProductImageRepository productImageRepository;
+    // 공동구매, 카트 등 나중에 생기면 얘네도 delete할때 먼저 지워야 함.
 
     @Override
     public List<ProductDto> getAllProducts() {
@@ -44,18 +48,28 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public ProductDto updateProduct(Long id, ProductDto dto) {
+    public void updateProduct(Long id, ProductDto dto) {
         Product existing = repository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Product not found"));
 
         Product updated = dto.toEntity();
         updated.setId(existing.getId());
 
-        return ProductDto.fromEntity(repository.save(updated));
+        ProductDto.fromEntity(repository.save(updated));
     }
 
-    @Override
+    @Transactional
     public void deleteProduct(Long id) {
-        repository.deleteById(id);
+        Optional<Product> product = repository.findById(id);
+        if (product.isPresent()) {
+            System.out.println(">> 삭제 대상 존재, product = " + product.get());
+            // 자식 엔티티 먼저 삭제
+            productImageRepository.deleteAllByProductId(id);
+
+            // 부모 엔티티 삭제
+            repository.deleteById(id);
+        } else {
+            System.out.println(">> 삭제 대상 없음, id = " + id);
+        }
     }
 }
